move token balance functionality to transaction-status crate
