+++ runtime/Cargo.toml
@@ -35,7 +35,6 @@ rayon = "1.4.1"
 regex = "1.3.9"
 serde = { version = "1.0.112", features = ["rc"] }
 serde_derive = "1.0.103"
-solana-account-decoder = { path = "../account-decoder", version = "1.5.0" }
 solana-config-program = { path = "../programs/config", version = "1.5.0" }
 solana-frozen-abi = { path = "../frozen-abi", version = "1.5.0" }
 solana-frozen-abi-macro = { path = "../frozen-abi/macro", version = "1.5.0" }
@@ -47,7 +46,6 @@ solana-sdk = { path = "../sdk", version = "1.5.0" }
 solana-stake-program = { path = "../programs/stake", version = "1.5.0" }
 solana-vote-program = { path = "../programs/vote", version = "1.5.0" }
 solana-secp256k1-program = { path = "../programs/secp256k1", version = "1.5.0" }
-spl-token-v2-0 = { package = "spl-token", version = "=2.0.8" }
 symlink = "0.1.0"
 tar = "0.4.28"
 tempfile = "3.1.0"
