+++ runtime/src/lib.rs
@@ -27,12 +27,10 @@ pub mod snapshot_utils;
 pub mod stakes;
 pub mod status_cache;
 mod system_instruction_processor;
-pub mod token_balances;
 pub mod transaction_batch;
 pub mod transaction_utils;
 pub mod vote_sender_types;
 
-extern crate solana_account_decoder;
 extern crate solana_config_program;
 extern crate solana_stake_program;
 extern crate solana_vote_program;
