+++ sys-tuner/src/main.rs
@@ -151,7 +151,8 @@ fn main() {
             info!("Tuning the system now");
             #[cfg(target_os = "linux")]
             {
-                // tune_poh_service_priority(peer_uid);
+                [cfg(target_os = "linux")]
+            { // tune_poh_service_priority(peer_uid);
             }
         }
     }
